Updated and performance improved, more info in desc
Code simplification and hash substitution improved

The hashes were not always substituted in a good way, depending on
whether there was a comment after the old hash or not. Now this is
fixed.Custom environment can now be added and the script will check if
thepath exits, this makes the script work even with the new
environments.The scripts will now look for the correct line. For example
it looks for '/adl:' instead of just 'adl:' (which was causing
errors).Some cases have been changed and grouped by the action they need
to do.